+++ HISTORY.md
@@ -1,6 +1,6 @@
 # Crass Changelog
 
-## 1.0.7 (unreleased)
+## 1.0.7 (2026-06-25)
 
 ### Security